rename `platform-client` -> `platform-rpc`
