+++ packages/platform-rpc/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@anticrm/platform-client",
-  "version": "0.1.1",
+  "name": "@anticrm/platform-rpc",
+  "version": "0.1.0",
   "main": "src/index.ts",
   "author": "Anticrm Platform Contributors (https://github.com/anticrm)",
   "license": "EPL-2.0",
